+++ src/manage_sql.c
@@ -48166,7 +48166,10 @@ manage_empty_trashcan ()
 
   /* Remove report formats last, because dir deletion can't be rolled back. */
   if (empty_trashcan_report_formats ())
-    return -1;
+    {
+      sql_rollback ();
+      return -1;
+    }
 
   sql_commit ();
   return 0;
@@ -51941,12 +51944,13 @@ manage_delete_user (GSList *log_config, const gchar *database,
  * @param[in]  log_config  Log configuration.
  * @param[in]  database    Location of manage database.
  * @param[in]  role_name   Role name.
+ * @param[in]  verbose     Whether to print UUID.
  *
  * @return 0 success, -1 error.
  */
 int
 manage_get_users (GSList *log_config, const gchar *database,
-                  const gchar* role_name)
+                  const gchar* role_name, int verbose)
 {
   iterator_t users;
   int ret;
@@ -51973,15 +51977,19 @@ manage_get_users (GSList *log_config, const gchar *database,
           return -1;
         }
       init_iterator (&users,
-                     "SELECT name FROM users"
+                     "SELECT name, uuid FROM users"
                      " WHERE id IN (SELECT \"user\" FROM role_users"
                      "              WHERE role = %llu);",
                      role);
     }
   else
-    init_iterator (&users, "SELECT name FROM users;");
+    init_iterator (&users, "SELECT name, uuid FROM users;");
   while (next (&users))
-    printf ("%s\n", iterator_string (&users, 0));
+    if (verbose)
+      printf ("%s %s\n", iterator_string (&users, 0), iterator_string (&users, 1));
+    else
+      printf ("%s\n", iterator_string (&users, 0));
+
   cleanup_iterator (&users);
 
   manage_option_cleanup ();